Update Magdwick benchmark.
Magdwick benchmark can now use the uncertainty aware instructions.
