+++ benchmarks/source/riscv_madgwick/.gitignore
@@ -1,0 +1,1 @@
+*.s
