+++ benchmarks/source/riscv_madgwick/init.S
@@ -1,0 +1,16 @@
+.globl _start
+.globl __errno
+.align	4
+
+_start:
+init:
+.option push
+.option norelax
+	la		gp, __global_pointer$;
+.option pop
+	call	main;
+	call 	exit;
+
+
+__errno:
+	.long	0
