+++ benchmarks/source/riscv_madgwick/run.m
@@ -1,0 +1,7 @@
+sizemem			98304
+settimerdelay		40000
+clockintr		1
+cacheoff
+srecl			main.sr
+run
+on
