--- benchmarks/source/riscv_magwick/.gitignore
@@ -1,1 +1,0 @@
-*.s
