--- benchmarks/source/riscv_magwick/run.m
@@ -1,7 +1,0 @@
-sizemem			131072
-settimerdelay		40000
-clockintr		1
-cacheoff
-srecl			main.sr
-run
-on
